pip_requirements.txt
pytelegrambotapi==3.7.2
psycopg2-binary==2.8.5
Pillow==7.2.0
Openpyxl==3.0.5